requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
